+++ internal/api/handler.go
@@ -2,13 +2,13 @@ package api
 
 import (
 	"context"
-	"errors"
 	"net/http"
 	"strconv"
 	"time"
 
 	"github.com/rs/zerolog/log"
 
+	json "github.com/json-iterator/go"
 	"github.com/s-hammon/volta/internal/api/models"
 	"github.com/s-hammon/volta/pkg/hl7"
 )
@@ -29,7 +29,7 @@ type API struct {
 	Client HealthcareClient
 }
 
-func New(db Repository, client HealthcareClient) *http.ServeMux {
+func New(db Repository, client HealthcareClient) http.Handler {
 	a := &API{
 		DB:     db,
 		Client: client,
@@ -37,75 +37,72 @@ func New(db Repository, client HealthcareClient) *http.ServeMux {
 
 	mux := http.NewServeMux()
 
-	// TODO: possiby add a health check?
-
 	mux.HandleFunc("POST /", a.handleMessage)
+	mux.HandleFunc("GET /healthz", handleReadiness)
 
-	return mux
+	loggedMux := middlwareLogging(mux)
+
+	return loggedMux
 }
 
 func (a *API) handleMessage(w http.ResponseWriter, r *http.Request) {
-	// TODO: reimplement logging (use middleware, perhaps)
-	// time this function
 	start := time.Now()
 	var logMsg logMsg
 
 	if r.Body == nil {
-		logMsg.Error(errors.New("empty request body"), "", "")
+		logMsg.Result = "empty request body"
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
 	defer r.Body.Close()
-	logMsg.notifSize = r.Header.Get("Content-Length")
+	logMsg.NotifSize = r.Header.Get("Content-Length")
 
 	m, err := NewPubSubMessage(r.Body)
 	if err != nil {
-		logMsg.Error(err, "", "")
+		logMsg.Result = err.Error()
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
 
 	msgMap, err := a.Client.GetHL7V2Message(string(m.Message.Data))
 	if err != nil {
-		logMsg.Error(err, "", "")
+		logMsg.Result = err.Error()
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
-	logMsg.hl7Size = strconv.Itoa(len(msgMap))
+	logMsg.Hl7Size = strconv.Itoa(len(msgMap))
 
 	switch m.Message.Attributes.Type {
 	case "ORM":
-		// TODO: to interface
 		orm := models.ORM{}
 		if err = hl7.Unmarshal(msgMap, &orm); err != nil {
-			logMsg.Error(err, orm.MSH.SendingFac, orm.MSH.ControlID)
+			logMsg.Result = err.Error()
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
 
 		if err = a.DB.UpsertORM(context.Background(), orm); err != nil {
-			logMsg.Error(err, orm.MSH.SendingFac, orm.MSH.ControlID)
+			logMsg.Result = err.Error()
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
-		logMsg.result = "ORM processed successfully"
+		logMsg.Result = "ORM processed successfully"
 
 	case "ORU":
-		// TODO: to interface
 		oru := models.ORU{}
 		if err = hl7.Unmarshal(msgMap, &oru); err != nil {
-			logMsg.Error(err, oru.MSH.SendingFac, oru.MSH.ControlID)
+			logMsg.Result = err.Error()
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
 
 		if err = a.DB.InsertORU(context.Background(), oru); err != nil {
-			logMsg.Error(err, oru.MSH.SendingFac, oru.MSH.ControlID)
+			logMsg.Result = err.Error()
 			w.WriteHeader(http.StatusInternalServerError)
 			return
 		}
 
-		logMsg.result = "ORU processed successfully"
+		logMsg.Result = "ORU processed successfully"
 	case "ADT":
 		log.Warn().
 			Str("messagePath", string(m.Message.Data)).
@@ -113,13 +110,20 @@ func (a *API) handleMessage(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusNotImplemented)
 		return
 	default:
-		logMsg.Error(errors.New("unknown message type"), "", "")
+		logMsg.Result = "unsupported message type"
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
 
-	logMsg.elapsed = time.Since(start)
-	logMsg.Log(logMsg.result)
+	logMsg.Elapsed = time.Since(start).Seconds() * 1000 // milliseconds
 
+	logBytes, err := json.Marshal(logMsg)
+	if err != nil {
+		log.Error().Err(err).Msg("could not marshal log message")
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusCreated)
+	w.Write(logBytes)
 }